+++ doc/CHANGES.rst
@@ -21,6 +21,8 @@ Future
 * [IMP] ModelAttribute used to declarate the need of specific attribute and
   get the attribute or the foreign key from this attribute
 * [IMP] ModelAttributeAdapter, get a ModelAttribute from ModelAttribute or str
+* [IMP] ModelRepr, Speudo representation of a Model
+* [IMP] ModelAdapter, get a ModelRepr from ModelRepr or str
 
 0.4.1
 -----